Changed Extend To Use Recursion
Using recursion and prototype methods, combining things i've learned, I
reimplemented extend. This takes the last item's keypairs, and adds it
the next to last item's keypairs. Until it reaches our base case.
The key thing to understand is how to use call and apply. I place the
focus on arguments but i use the slice for []. Then I use the call but
I don't need to use the thisfocus can be anything.